Changed int to bigint

+++ db/migrate/20200512182502_create_positions.rb
@@ -1,7 +1,7 @@
 class CreatePositions < ActiveRecord::Migration[5.2]
   def change
     create_table :positions do |t|
-      t.int :code
+      t.bigint :code
       t.string :name
       t.date :date_debut_position
       t.date :date_fin_position
