+++ db/migrate/20200512191228_create_expatries.rb
@@ -1,7 +1,7 @@
 class CreateExpatries < ActiveRecord::Migration[5.2]
   def change
     create_table :expatries do |t|
-      t.int :code
+      t.bigint :code
       t.string :name
 
       t.timestamps
