+++ db/migrate/20200525183420_add_code_to_expatries.rb
@@ -1,5 +1,5 @@
 class AddCodeToExpatries < ActiveRecord::Migration[5.2]
   def change
-    add_column :expatries, :code, :integer
+    add_column :expatries, :code, :bigint
   end
 end
